requirements.txt
# Base Flask Requirements
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0

# Google Cloud Dialogflow
google-cloud-dialogflow==2.14.0
protobuf==3.19.6

# Data Analysis & Machine Learning
numpy==1.23.5  # Diubah ke versi yang stabil
pandas==1.2.4
scikit-learn==0.24.2
pillow==9.5.0  # Menambahkan Pillow dengan versi spesifik
matplotlib==3.4.3
seaborn==0.11.2
folium==0.12.1
deap==1.3.1
tensorflow==2.8.0

# Computer Vision
opencv-python==4.5.3.56